requirements.txt
streamlit
python-dotenv
google-adk==1.14.0
langchain-community==0.3.27
google-cloud-aiplatform
